requirements.txt
python-dotenv
wikipedia
openai
selenium
google-api-python-client
oauth2client